LXC Version 1.0 updates
In LXC version 1.0 it seems that the rootfs is passed into the template
upon creation, as such I have added support so that the template can work
with LXC 1.0.1 and earlier.

Other updates add hard link support when creating the distro as well as
installing the english language pack and setting other distro support
to raise a notimplemented exception until I get around to implementing it.